moved kube to actions folder
