+++ lua/kubectl/actions/kube.lua
