+++ lua/kubectl/init.lua
@@ -1,7 +1,7 @@
 local async = require("plenary.async")
 local config = require("kubectl.config")
 local hl = require("kubectl.actions.highlight")
-local kube = require("kubectl.utils.kube")
+local kube = require("kubectl.actions.kube")
 local mappings = require("kubectl.mappings")
 local pod_view = require("kubectl.views.pods")
 local state = require("kubectl.utils.state")
