+++ lua/kubectl/mappings.lua
@@ -2,7 +2,7 @@ local config = require("kubectl.config")
 local configmaps_view = require("kubectl.views.configmaps")
 local deployments_view = require("kubectl.views.deployments")
 local filter_view = require("kubectl.views.filter")
-local kube = require("kubectl.utils.kube")
+local kube = require("kubectl.actions.kube")
 local namespace_view = require("kubectl.views.namespace")
 local pods_view = require("kubectl.views.pods")
 local secrets_view = require("kubectl.views.secrets")
